Triangle intersection function, sphere intersection function

+++ Ray Tracer/SceneObject.h
@@ -5,12 +5,11 @@ public:
 	CSceneObject(void);
 	~CSceneObject(void);
 
-	int reflect;
-	int shininess;
+	float reflect;
+	float shininess;
 	glm::vec3 amb;
 	glm::vec3 diff;
 	glm::vec3 spec;
-
 
 	virtual float intersect( CRay* ray ) = 0;
 };
